+++ WindBoard/Features/Export/Models/BoardPdfExportOptions.cs
@@ -1,8 +1,7 @@
-namespace WindBoard.Exporting
+namespace WindBoard.Features.Export.Models
 {
     /// <summary>
     /// PDF 导出参数。
     /// </summary>
     internal sealed record BoardPdfExportOptions(BoardRasterExportOptions RasterOptions);
 }
-
